Move HintProps to types
