+++ src/components/hint/index.tsx
@@ -5,13 +5,7 @@ import {
   StyleSheet,
   AccessibilityInfo,
   findNodeHandle,
-  GestureResponderEvent,
-  ImageSourcePropType,
   TouchableWithoutFeedback,
-  ImageStyle,
-  StyleProp,
-  TextStyle,
-  ViewStyle,
   View as RNView
 } from 'react-native';
 import {Typography, Spacings, Colors, BorderRadiuses, Shadows} from '../../style';
@@ -21,15 +15,7 @@ import Text from '../text';
 import Image from '../image';
 import Modal from '../modal';
 import TouchableOpacity from '../touchableOpacity';
-import {
-  ContentType,
-  HintPositions,
-  HintPositionStyle,
-  HintTargetFrame,
-  Paddings,
-  Position,
-  TARGET_POSITIONS
-} from './types';
+import {HintPositions, HintPositionStyle, HintProps, Paddings, Position, TARGET_POSITIONS} from './types';
 
 import {useDidUpdate} from 'hooks';
 import useHintAnimation from './hooks/useHintAnimation';
@@ -42,98 +28,6 @@ const DEFAULT_COLOR = Colors.$backgroundPrimaryHeavy;
 const DEFAULT_HINT_OFFSET = Spacings.s4;
 const DEFAULT_EDGE_MARGINS = Spacings.s5;
 const HINT_MIN_WIDTH = 68;
-
-export interface HintProps {
-  /**
-   * Control the visibility of the hint
-   */
-  visible?: boolean;
-  /**
-   * The hint background color
-   */
-  color?: string;
-  /**
-   * The hint message
-   */
-  message?: ContentType | ContentType[];
-  /**
-   * The hint message custom style
-   */
-  messageStyle?: StyleProp<TextStyle>;
-  /**
-   * Icon to show next to the hint's message
-   */
-  icon?: ImageSourcePropType;
-  /**
-   * The icon's style
-   */
-  iconStyle?: StyleProp<ImageStyle>;
-  /**
-   * The hint's position
-   */
-  position?: HintPositions;
-  /**
-   * Provide custom target position instead of wrapping a child
-   */
-  targetFrame?: HintTargetFrame;
-  /**
-   * Open the hint using a Modal component
-   */
-  useModal?: boolean;
-  /**
-   * Show side tips instead of the middle tip
-   */
-  useSideTip?: boolean;
-  /**
-   * The hint's border radius
-   */
-  borderRadius?: number;
-  /**
-   * Hint margins from screen edges
-   */
-  edgeMargins?: number;
-  /**
-   * Hint offset from target
-   */
-  offset?: number;
-  /**
-   * Callback for Hint press
-   */
-  onPress?: () => void;
-  /**
-   * Callback for the background press
-   */
-  onBackgroundPress?: (event: GestureResponderEvent) => void;
-  /**
-   * Color for background overlay (require onBackgroundPress)
-   */
-  backdropColor?: string;
-  /**
-   * The hint container width
-   */
-  containerWidth?: number;
-  /**
-   * Custom content element to render inside the hint container
-   */
-  customContent?: JSX.Element;
-  /**
-   * Remove all hint's paddings
-   */
-  removePaddings?: boolean;
-  /**
-   * Enable shadow (for hint with white background only)
-   */
-  enableShadow?: boolean;
-  /**
-   * The hint's test identifier
-   */
-  testID?: string;
-  /**
-   * Additional styling
-   */
-  style?: StyleProp<ViewStyle>;
-  children?: React.ReactNode;
-}
 
 const NewHint = (props: HintProps) => {
   const {
